pyproject.toml
[project]
name = "async_chat"
version = "0.1.0"
description = ""
authors = [
    {name = "xifOO",email = "wedyi28111@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.32.0,<0.33.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "ruff (>=0.12.12,<0.13.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = [
  "E4", "E7", "E9",
  "F",
  "I",
  "B",
  "A",
  "C4",
  "S",
]
ignore = ["B904"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

per-file-ignores = { "tests/*" = ["S101"] }

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"